+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Added a public function `Set-TaskScriptVariables` that is used to dot-source
+- Added a public function `Set-TaskScriptVariable` that is used to dot-source
   task variables for re-use over multiple build tasks.
 
 ### Fixed